Refactor code to be more modular. open_cal() now returns both the list of events as well as the calendar itself, so that you can get the field calendar.CALENDAR_TIMEZONE
This is useful in formatting events that happen for the entire day, as
their start and end do not have start fields.

Also implemented format_events so that all events start and end time are
in datetime.datetime format, rather than datetime.date format. This
allows sortedevents to run regardless of events are for a specified time
or the whole day.